Added Sound2D and SoundManger classes, bug fixes
Sounds can now be created through the SoundManager class and played on
their own.

Bug fixes (compilation errors) in player and gameplayHandler classes,
see DEBUG comments for spots and what the fix was.

+++ Code/Model/Coord.hpp
@@ -5,7 +5,8 @@
 
 namespace Model
 {
-	typedef Helper::Point2i Coord;
+	typedef Helper::Point2f Coord;  // DEBUG: changed from Point2i to Point2f because of type clash 
+									// in method call: UpdateMovement in GameplayHandler.
 	//struct Coord
 	//{
 	//	Coord();
